[WIP] Add "Stream" Encoder
This commit introduces the StreamEncoder type, which provides access
to the low-level event-based interface, wrapped in an easily-usable
form.

The StreamEncoder can be used by implementing the
`StreamMarshalYAML(*StreamEncoder) error` method on an type,
just like one might implement `MarshalYAML` on an type.

It can also be directly created using the `NewStreamEncoder()`
function (and the output can be retrieved using the `Finish()`
function).

The `StreamEncoder` is then used by "emitting" values according to
what's expected:

- to emit a value as the encoder sees fit, use the `EmitValue` function
- to produce a sequence, use `BeginSequence`, followed by a value for
  each item you want, followed by `EndSequence`
- to produce a mapping, use `BeginMapping`, followed by pairs of items
  (key first, then value) for each key/value pair you want, followed
  by `EndSequence`

To facilitate the process of encoding a struct or map on a pair-by-pair
basis, the function `EmitAsMap` is provided.  It takes a struct,
map, or MapSlice and calls the given function for each key/value
pair that would have been emitted.  The function can the emit a
key/value pair (either the same, or different), or not emit anything.