[WIP] Add "Stream" Decoder
This commit adds the StreamDecoder type, which provides access to
the low-level parsing event stream, wrapped in an easily-usable
form.

A new StreamDecoder can be created using the `NewStreamDecoder`
function.  Then, the user can call `NextEvent` to get the next
event in the event stream.  This continues until a `FinishEvent`
occurs, at which point the stream is finished and the decoder no
longer functions.