[WIP] Add Comment-Parsing Support
This commit adds support for parsing comments.  Comments may will
not be parsed by the normal unmarshaller; comments will appear only
as `CommentEvent` when using the `StreamDecoder`.  Each `CommentEvent`
has a value of the text of the comment.

This also adds in a round-trip test for the stream parsers, to ensure
that everything is preserved parsed and encoded properly.