Add support for aligning comments to a column
This commit adds support for aligning same-line comments to a
particular column.  Specifically, it causes same-line comments
to start at minimum at that particular column.

It also exposes a new object, Marshaller, which allows a user
to control whether this occurs, as well as the desired wrap-width
and indent size.